chore(pkg/client): move unit testing to integration package to avoid circular references
