+++ pkg/integration/signature_verifier_interceptor_test.go
@@ -1,21 +1,38 @@
-package client
+/*
+Copyright 2021 CodeNotary, Inc. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package integration
 
 import (
 	"context"
+	"testing"
+
 	"github.com/codenotary/immudb/pkg/api/schema"
+	ic "github.com/codenotary/immudb/pkg/client"
 	"github.com/codenotary/immudb/pkg/server"
 	"github.com/codenotary/immudb/pkg/signer"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
-	"testing"
 )
 
 func TestSignatureVerifierInterceptor(t *testing.T) {
 
 	pk, err := signer.ParsePublicKeyFile("./../../test/signer/ec1.pub")
 	require.NoError(t, err)
-	c := DefaultClient().WithServerSigningPubKey(pk)
+	c := ic.DefaultClient().WithServerSigningPubKey(pk)
 
 	// creation and state sign
 	state := &schema.ImmutableState{
@@ -41,7 +58,7 @@ func TestSignatureVerifierInterceptor(t *testing.T) {
 func TestSignatureVerifierInterceptorUnableToVerify(t *testing.T) {
 	pk, err := signer.ParsePublicKeyFile("./../../test/signer/ec1.pub")
 	require.NoError(t, err)
-	c := DefaultClient().WithServerSigningPubKey(pk)
+	c := ic.DefaultClient().WithServerSigningPubKey(pk)
 
 	// creation and state sign
 	state := &schema.ImmutableState{
@@ -62,7 +79,7 @@ func TestSignatureVerifierInterceptorUnableToVerify(t *testing.T) {
 func TestSignatureVerifierInterceptorSignatureDoesntMatch(t *testing.T) {
 	pk, err := signer.ParsePublicKeyFile("./../../test/signer/ec1.pub")
 	require.NoError(t, err)
-	c := DefaultClient().WithServerSigningPubKey(pk)
+	c := ic.DefaultClient().WithServerSigningPubKey(pk)
 
 	// creation and state sign
 	state := &schema.ImmutableState{
@@ -85,7 +102,7 @@ func TestSignatureVerifierInterceptorSignatureDoesntMatch(t *testing.T) {
 }
 
 func TestSignatureVerifierInterceptorNoPublicKey(t *testing.T) {
-	c := DefaultClient().WithServerSigningPubKey(nil)
+	c := ic.DefaultClient().WithServerSigningPubKey(nil)
 	// creation and state sign
 	state := &schema.ImmutableState{
 		TxId:   0,
