+++ pkg/integration/streams_zscan_and_history_test.go
@@ -1,4 +1,19 @@
-package client
+/*
+Copyright 2021 CodeNotary, Inc. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package integration
 
 import (
 	"bufio"
@@ -9,6 +24,7 @@ import (
 	"testing"
 
 	"github.com/codenotary/immudb/pkg/api/schema"
+	ic "github.com/codenotary/immudb/pkg/client"
 	"github.com/codenotary/immudb/pkg/server"
 	"github.com/codenotary/immudb/pkg/server/servertest"
 	"github.com/codenotary/immudb/pkg/stream"
@@ -27,7 +43,7 @@ func TestImmuClient_StreamZScan(t *testing.T) {
 	bs.Start()
 	defer bs.Stop()
 
-	client, err := NewImmuClient(DefaultOptions().WithDialOptions(
+	client, err := ic.NewImmuClient(ic.DefaultOptions().WithDialOptions(
 		&[]grpc.DialOption{grpc.WithContextDialer(bs.Dialer), grpc.WithInsecure()},
 	))
 	require.NoError(t, err)
@@ -85,7 +101,7 @@ func TestImmuClient_StreamHistory(t *testing.T) {
 	bs.Start()
 	defer bs.Stop()
 
-	client, err := NewImmuClient(DefaultOptions().WithDialOptions(
+	client, err := ic.NewImmuClient(ic.DefaultOptions().WithDialOptions(
 		&[]grpc.DialOption{grpc.WithContextDialer(bs.Dialer), grpc.WithInsecure()},
 	))
 	require.NoError(t, err)
